+++ server/api/auth/token.get.ts
@@ -1,0 +1,3 @@
+import { getAuthTokenStatus } from '~~/server/services/authTokenService'
+
+export default defineEventHandler(() => getAuthTokenStatus())